remove errts and fitts
use inputs to select subs

+++ blind/clear_blind_disk.tcsh
@@ -1,19 +1,29 @@
 #! /bin/csh
-foreach ub (`count -dig 2 17 19`)
+# foreach ub (`count -dig 2 17 19`)
+# from the first to the second input
+foreach ub (`count -dig 2 $1 $2`)
 
 set sub = S${ub}
 set datafolder=/Volumes/WD_F/gufei/blind/${sub}/
-cd "${datafolder}"
-# make dir to save masks
-set analysis=pade
-set subj = ${sub}.${analysis}
-cd ${subj}.results
+# if datafolder exist
+if ( -d "${datafolder}" ) then
+    cd "${datafolder}"
 
-# find volreg number
-@ pbvol = `ls pb0?.*.r01.volreg+orig.HEAD | cut -d . -f1 | cut -c3-4` - 1
+    # make dir to save masks
+    set analysis=pade
+    set subj = ${sub}.${analysis}
+    cd ${subj}.results
 
-# remove files
-rm all_runs*
-rm pb0[0-${pbvol}]*
+    # find volreg number
+    @ pbvol = `ls pb0?.*.r01.volreg+orig.HEAD | cut -d . -f1 | cut -c3-4` - 1
+
+    # remove files
+    rm all_runs*
+    rm pb0[0-${pbvol}]*
+    rm errts*
+    rm fitts*
+else
+    echo "${datafolder} not exsist"
+endif
 
 end
